retry provider, cluster processing pseudo, helpers
