+++ anchor/eth/src/lib.rs
@@ -2,6 +2,6 @@ pub use sync::SsvEventSyncer;
 mod event_parser;
 mod event_processor;
 mod gen;
-mod sync;
 mod network_actions;
+mod sync;
 mod util;
